backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
python-magic
PyPDF2
docx2txt
transformers
langchain
langchain-huggingface
torch
